:recycle: sample_lib as local and separate docstrings tests
