+++ .isort.cfg
@@ -7,4 +7,4 @@ import_heading_thirdparty=Third Party
 import_heading_firstparty=First Party
 import_heading_localfolder=Local
 known_firstparty=alog,aconfig,jtd_to_proto,import_tracker
-known_localfolder=caikit,tests
+known_localfolder=caikit,sample_lib,tests
