+++ tests/core/data_model/streams/test_data_stream.py
@@ -15,13 +15,11 @@
 # Standard
 import os
 
-# Third Party
-from sample_lib.blocks.sample_task.sample_implementation import SampleBlock
-from sample_lib.data_model.sample import SampleInputType, SampleOutputType
-
 # Local
 from caikit.core import data_model as core_dm
 from caikit.core.augmentors import AugmentorBase
+from sample_lib.blocks.sample_task.sample_implementation import SampleBlock
+from sample_lib.data_model.sample import SampleInputType, SampleOutputType
 
 # Unit Test Infrastructure
 from tests.base import TestCaseBase
